+++ diskover_dupes.py
@@ -48,14 +48,14 @@ def index_dupes(hashgroup, cliargs):
         index_bulk_add(es, file_id_list, config, cliargs)
 
 
-def start_file_threads(file_in_thread_q, file_out_thread_q):
+def start_file_threads():
     for i in range(config['dupes_threads']):
-        thread = Thread(target=md5_hasher, args=(file_in_thread_q, file_out_thread_q,))
+        thread = Thread(target=md5_hasher)
         thread.daemon = True
         thread.start()
 
 
-def md5_hasher(file_in_thread_q, file_out_thread_q):
+def md5_hasher():
     while True:
         item = file_in_thread_q.get()
         filename, atime, mtime, cliargs = item
@@ -179,10 +179,6 @@ def verify_dupes(hashgroup, cliargs):
 
     # run md5 sum check if bytes were same
     hashgroup_md5 = {}
-    # set up python Queue for threaded file md5 checking
-    file_in_thread_q = pyQueue()
-    file_out_thread_q = pyQueue()
-    start_file_threads(file_in_thread_q, file_out_thread_q)
 
     # do md5 check on files with same byte hashes
     for key, value in list(hashgroup_bytes.items()):
@@ -331,3 +327,9 @@ def dupes_finder(es, q, cliargs, logger):
 
     if bar:
         bar.finish()
+
+
+# set up python Queue for threaded file md5 checking
+file_in_thread_q = pyQueue()
+file_out_thread_q = pyQueue()
+start_file_threads()